Removed Unnecessary Files & Updated API and Checks
remember to change the path to "register" from "register2"
