+++ src/java/models/User.java
@@ -1,30 +1,22 @@
 package models;
 
+
 public class User {
-    
     private String username;
     private String password;
-    private String firstname;
-    private String lastname;
+    private String firstName;
+    private String lastName;
+    private Address address;
+    private Veterinarian veterinarian; // Unused at the time
     private String email;
-    
-    public User()
-    {
-        
-    }
-    
-    public User(String username, String password, String firstname, String lastname, String email)
-    {
-        this.username = username;
-        this.password = password;
-        this.firstname = firstname;
-        this.lastname = lastname;
-        this.email = email;
-    }
+    private String phoneNumber;
+    private String emergencyPhone;
+    private String emergencyName;
+    private boolean isActive;
+    private boolean isDisabled;
 
-    public User(String username, String password) {
-        this.username = username;
-        this.password = password;
+    public User() {
+        this.address = new Address();
     }
 
     public String getUsername() {
@@ -43,20 +35,36 @@ public class User {
         this.password = password;
     }
 
-    public String getFirstname() {
-        return firstname;
+    public String getFirstName() {
+        return firstName;
     }
 
-    public void setFirstname(String firstname) {
-        this.firstname = firstname;
+    public void setFirstName(String firstName) {
+        this.firstName = firstName;
     }
 
-    public String getLastname() {
-        return lastname;
+    public String getLastName() {
+        return lastName;
     }
 
-    public void setLastname(String lastname) {
-        this.lastname = lastname;
+    public void setLastName(String lastName) {
+        this.lastName = lastName;
+    }
+
+    public Address getAddress() {
+        return address;
+    }
+
+    public void setAddress(Address address) {
+        this.address = address;
+    }
+
+    public Veterinarian getVeterinarian() {
+        return veterinarian;
+    }
+
+    public void setVeterinarian(Veterinarian veterinarian) {
+        this.veterinarian = veterinarian;
     }
 
     public String getEmail() {
@@ -66,10 +74,44 @@ public class User {
     public void setEmail(String email) {
         this.email = email;
     }
-    
-    @Override
-    public String toString()
-    {
-        return "Firstname: " + firstname + " Lastname: " + lastname;
+
+    public String getPhoneNumber() {
+        return phoneNumber;
+    }
+
+    public void setPhoneNumber(String phoneNumber) {
+        this.phoneNumber = phoneNumber;
+    }
+
+    public String getEmergencyPhone() {
+        return emergencyPhone;
+    }
+
+    public void setEmergencyPhone(String emergencyPhone) {
+        this.emergencyPhone = emergencyPhone;
+    }
+
+    public String getEmergencyName() {
+        return emergencyName;
+    }
+
+    public void setEmergencyName(String emergencyName) {
+        this.emergencyName = emergencyName;
+    }
+
+    public boolean isIsActive() {
+        return isActive;
+    }
+
+    public void setIsActive(boolean isActive) {
+        this.isActive = isActive;
+    }
+
+    public boolean isIsDisabled() {
+        return isDisabled;
+    }
+
+    public void setIsDisabled(boolean isDisabled) {
+        this.isDisabled = isDisabled;
     }
 }
